Atualizando a Documentação

+++ src/controller/register/RegisterController.ts
@@ -1,7 +1,7 @@
-import { Json } from "aws-sdk/clients/robomaker";
 import { Request, Response } from "express";
-import {Sqs} from "../../config/SQS/Sqs";
-import {RegisterUseCases} from "./RegisterUsecases"
+import { AppError } from "../../errors/AppError";
+import { Sqs } from "../../config/SQS/Sqs";
+import { RegisterUseCases } from "./RegisterUsecases"
 
 class RegisterController{
   async handle(request:Request, response: Response): Promise<Response>{
@@ -9,7 +9,7 @@ class RegisterController{
     const  sqs = new Sqs();
     const registerUsecases = new RegisterUseCases()
     const { employeeId, employerId, includedAt} = request.body;
-    const wait = ms => new Promise(resolve => setTimeout(resolve, ms));
+    const wait = (ms: number | undefined) => new Promise(resolve => setTimeout(resolve, ms));
 
     //ADD THE NEW DATA INTO THE QUEUE
     await sqs.sendCommand(employeeId, employerId, includedAt)
@@ -21,29 +21,37 @@ class RegisterController{
     const QueueSize = await sqs.QueueSizeCommand()
     // console.log('size', QueueSize)
 
+    //INITIALIZE AN EMPTY ARRAY
     let finalResponse : JSON [] = []
 
+    /* FOR ALL ELEMENT IN THE QUEUE 
+        Get the first element in the Queue
+        Resolve the pendency 
+        Delete it when is Done.
+    */
     for (let i=0;i<QueueSize;i++){
 
       const element = await sqs.receiveCommand()
 
-      // console.log('element', element.Body)
+      try{
+        var results = await registerUsecases.execute(JSON.parse(element.Body))
+        
+        // console.log(' resultados ',results)
+        if(results.hasOwnProperty('message') && results.hasOwnProperty('system')){
 
-      // console.log(JSON.parse(element.Body))
-      var results = await registerUsecases.execute(JSON.parse(element.Body))
+          //DELETED AN ELEMENT OF THE QUEUE
+          let _= await sqs.deleteCommand(element.ReceiptHandle)
 
+          // ADD TO THE ARRAYS ALL RETURN VALUES OF ELEMENTS EXECUTIONS
+          finalResponse.push(results)
+        }
+      }catch(e){
+        throw new AppError("Element's Queue Failed !!!");
+      }
 
-      console.log(' resultados ',results)
-      // deleted element in the Queue
-      let _= await sqs.deleteCommand(element.ReceiptHandle)
-      finalResponse.push(results)
     }
-
-    // const result = registerUsecases.execute({includedAt,employeeId,employerId})
     return response.send(finalResponse);
-
   }
-
 }
 
 export {RegisterController}
